Scheduler fixed for time more than 4000 PWM software more resolution but with maximum 32mS Servo Enhanced more accurate and have more angles

+++ PWMSoftware.cpp
@@ -81,7 +81,7 @@ void PWMSoftware::setPeriod(uint16_t noofTicks){
 *_TCCRnB = tempTCCRnB;
 }
 
-void PWMSoftware::setPWM(uint16_t noofTicks){
+void PWMSoftware::setPWM(uint32_t noofTicks){
 *_OCRnX = noofTicks;
 }
