+++ PWMSoftware.h
@@ -9,7 +9,7 @@ class PWMSoftware
 	PWMSoftware();
 	void init(uint8_t timerNo, uint8_t ocrLetter);
 	void setPeriod(uint16_t noofTicks);
-	void setPWM(uint16_t noofTicks);
+	void setPWM(uint32_t noofTicks);
 	private:
 	volatile uint16_t *_OCRnX;
 	volatile uint8_t *_TCCRnA;
